Fix Elf32_Sym->st_info in elf modules
After compiling with mingw32-gcc, it is necessary to transform the
modules to the elf format with <objectcopy>.
It appears that in this translation the type of objects in the
structure Elf32_Sym-> st_info (STT_OBJECT, STT_FUNC ...) was lost.
To solve this I made the <PlayingBinaries> program that uses this
auxiliary configuration file.